Move PropertiesFileWriter to Plugin module
U-ways
