+++ propactive-plugin/src/main/kotlin/io/github/propactive/project/PropertiesFileWriter.kt
@@ -1,4 +1,4 @@
-package io.github.propactive.file
+package io.github.propactive.project
 
 import io.github.propactive.environment.EnvironmentModel
 import java.io.File
